create component to house notifications
