+++ resources/js/Layouts/AuthenticatedLayout.vue
@@ -8,8 +8,7 @@ import ResponsiveNavLink from '@/Components/ResponsiveNavLink.vue';
 import Toast from '@/Components/Misc/Toast.vue';
 import { Link, usePage } from '@inertiajs/vue3';
 import { currencies } from '@/currencies.js';
-import { useEchoNotification } from "@laravel/echo-vue";
-import Notification from '@/Components/Notifications/Notification.vue';
+import Notifications from '@/Components/Notifications/Notifications.vue';
 
 const props = defineProps({
     status: {
@@ -18,20 +17,6 @@ const props = defineProps({
     },
 });
 
-const notifications = ref(usePage().props.notifications);
-
-useEchoNotification(
-    `App.Models.User.${usePage().props.auth.user.id}`,
-    (notification) => {
-        // structure the notification the same way they are taken from db
-        const dataNotification = {
-            'data' : notification
-        };
-
-        notifications.value.push(dataNotification);
-    },
-);
- 
 // this will be part of the eventual exchange rework, choosing to show your balance in whichever currency
 // const currency = currencies.find((currency) => currency.code == usePage().props.auth.user.user_balance.currency);
 const user_balance = ref(usePage().props.auth.user.user_balance.amount);
@@ -184,29 +169,7 @@ const showingNavigationDropdown = ref(false);
                             </div>
 
                             <!-- Notifications -->
-                            <Dropdown align="right"
-                                width="80"
-                            >
-                                <template #trigger>
-                                    <button
-                                        type="button"
-                                        class="text-center inline-flex items-center rounded-md border border-transparent bg-white px-3 py-2 text-sm font-medium leading-4 text-gray-500 transition duration-150 ease-in-out hover:text-gray-700 focus:outline-none"
-                                    >
-                                        <i class="fa-solid fa-circle-exclamation text-2xl text-gray-400"></i>
-                                    </button>
-                                </template>
-                                <template #content v-if="notifications.length === 0">
-                                    <p class="block w-full px-4 py-2 text-start text-sm leading-5 text-gray-700 transition duration-150 ease-in-out hover:bg-gray-100 focus:bg-gray-100 focus:outline-none">
-                                        You have no new notifications
-                                    </p>
-                                </template>
-                                <template #content v-else>
-                                    <Notification v-for="notification in notifications"
-                                        :notification="notification"
-                                    >
-                                    </Notification>
-                                </template>
-                            </Dropdown>
+                            <Notifications />
                         </div>
                     </div>
                 </div>
